Fixes bug regarding Favorite filter
The bug occurred when the list was filtered to only show Favorites
and the last favorite was un-favorited.

This would leave the list invisible with no way for the user to
fix this.

A fail-safe has been added that prevents the favorite button from
being marked as "filter=true" when there are no favirotes.